save build time

+++ vte_script/dbuild.sh
@@ -45,6 +45,8 @@ declare -a unit_test_configs;
 declare -a rootfs_apd;
 declare -a gpu_branch;
 declare -a gpu_configs;
+declare -a target_configs;
+declare -a vte_target_configs;
 # As bash only support 1 dimension array below sequence is our assumption
 # 0   1  2  3  4  5  6  7  8  9  10 11
 #(23 25 28 31 35 37 50 50  51 53 53 61)
@@ -96,6 +98,8 @@ linux_libs_branch=("master" "master" "master" "master" "master" "master" "master
 #rootfs and vte apendix
 rootfs_apd=("" "" "" "" "" "" "" "" "" "" "" "" "" "" "" "" "" "" "");
 gpu_configs=("0" "0" "0" "0" "0" "0" "0" "0" "0" "0" "0" "1" "1" "1" "1" "1" "1" "1" "0");
+target_configs=("0" "0" "0" "0" "0" "0" "0" "0" "0" "0" "0" "1" "1" "1" "1" "1" "1" "1" "1");
+vte_target_configs=("0" "0" "0" "0" "0" "0" "0" "0" "0" "0" "0" "1" "1" "1" "1" "1" "1" "1" "1");
 
 branch_libs()
 {
@@ -111,6 +115,20 @@ branch_libs()
 make_libs()
 {
   iRC=0
+  if [ "$old_libs_config" = "${1}${2}" ]; then
+   if [ "old_libs_target" = "${3}${4}"  ]; then
+	echo "libs deployed already"
+        return $iRC
+   fi
+  sudo make DEST_DIR=${TARGET_ROOTFS}/imx${3}_rootfs${4} install -k || iRC=$(expr $iRC + 1)
+ 	if [ $deploy_target_rd -eq 1 ]; then
+  sudo make DEST_DIR=${TARGET_ROOTFS_RD}/imx${3}_rootfs${4} install -k || iRC=$(expr $iRC + 1)
+	fi
+     return $iRC
+  fi
+
+  old_libs_config=${1}${2}
+  old_libs_target=${3}${4}
   cd $LIB_DIR
   git add . 
   git commit -s -m"reset"
@@ -125,7 +143,9 @@ make_libs()
   make clean
   make PLATFORM=${2} CROSS_COMPILE=${TOOL_CHAIN}arm-none-linux-gnueabi- INCLUDE="-I${KERNEL_DIR}/include -I${KERNEL_DIR}/drivers/mxc/security/rng/include -I${KERNEL_DIR}/drivers/mxc/security/sahara2/include" -k || iRC=1
   sudo make DEST_DIR=${TARGET_ROOTFS}/imx${3}_rootfs${4} install -k || iRC=$(expr $iRC + 1)
+ 	if [ $deploy_target_rd -eq 1 ]; then
   sudo make DEST_DIR=${TARGET_ROOTFS_RD}/imx${3}_rootfs${4} install -k || iRC=$(expr $iRC + 1)
+	fi
   return $iRC
 }
 
@@ -141,8 +161,10 @@ deploy_firmware()
   if [ -e ${FIRMWARE_DIR}/firmware ]; then
     sudo rm -rf ${TARGET_ROOTFS}/imx${1}_rootfs${2}/lib/firmware
     sudo cp -af ${FIRMWARE_DIR}/firmware ${TARGET_ROOTFS}/imx${1}_rootfs${2}/lib/ || return 1
+ 	if [ $deploy_target_rd -eq 1 ]; then
     sudo rm -rf ${TARGET_ROOTFS_RD}/imx${1}_rootfs${2}/lib/firmware
     sudo cp -af ${FIRMWARE_DIR}/firmware ${TARGET_ROOTFS_RD}/imx${1}_rootfs${2}/lib/ || return 1
+ 	fi
   fi 
   return 0
 }
@@ -171,7 +193,9 @@ make_unit_test()
  fi
  sudo cp unit_test ${VTE_TARGET_PRE}/vte_mx${2}_${3}d/runtest/
  sudo cp unit_test ${VTE_TARGET_PRE2}/vte_mx${2}_${3}d/runtest/
+ if [ $deploy_vte_target_rd -eq 1 ]; then
  sudo cp unit_test ${VTE_TARGET_PRE3}/vte_mx${2}_${3}d/runtest/
+ fi
  old_ut_plat=$2
  old_ut_rc=0
  PLATFORM=$1
@@ -191,9 +215,11 @@ make_unit_test()
  DEPMOD=/bin/true INSTALL_MOD_PATH=${TARGET_ROOTFS}/imx${2}_rootfs${3} install -k || old_ut_rc=$(expr $old_ut_rc + 4)
  sudo  make PLATFORM=$PLATFORM DESTDIR=${TARGET_ROOTFS}/imx${2}_rootfs${3}/unit_tests \
  CROSS_COMPILE=${TOOL_CHAIN}arm-none-linux-gnueabi- install || old_ut_rc=$(expr $old_ut_rc + 8)
+ if [ $deploy_target_rd -eq 1 ]; then
  DEPMOD=/bin/true INSTALL_MOD_PATH=${TARGET_ROOTFS_RD}/imx${2}_rootfs${3} install -k || old_ut_rc=$(expr $old_ut_rc + 5)
  sudo  make PLATFORM=$PLATFORM DESTDIR=${TARGET_ROOTFS_RD}/imx${2}_rootfs${3}/unit_tests \
  CROSS_COMPILE=${TOOL_CHAIN}arm-none-linux-gnueabi- install || old_ut_rc=$(expr $old_ut_rc + 8)
+ fi
  return $old_ut_rc
 }
 
@@ -208,9 +234,11 @@ $UCONFDIR/u-config -s ${1}-config.txt u-boot-${1}-config.bin
 $UCONFDIR/u-config -s ${1}-config_rd.txt u-boot-${1}-config_rd.bin
 #rm -f ${1}_config.txt
 sudo cp u-boot-${1}-config.bin ${TARGET_ROOTFS}/imx${3}_rootfs${4}/root/u-boot-${1}-config.bin || return 3
+ if [ $deploy_target_rd -eq 1 ]; then
 sudo cp u-boot-${1}-config.bin ${TARGET_ROOTFS_RD}/imx${3}_rootfs${4}/root/u-boot-${1}-config_rd.bin || return 3
-rm -rf  u-boot-${1}-config.bin
 rm -rf  u-boot-${1}-config_rd.bin
+ fi
+rm -rf  u-boot-${1}-config.bin
 }
 
 make_uboot()
@@ -237,12 +265,19 @@ echo "make gpu driver $2 with $1"
 cd $GPU_DIR/driver
 if [ "$old_gpu_config" = $1 ];then
 if [ "$old_gpu_rc" -eq 0 ]; then
+  if [ $old_gpu_target = "${2}${3}" ]; then
+	echo "already deployed"
+	return $old_gpu_rc
+  fi
 sudo cp -a $GPU_DIR/build/sdk/drivers/* ${TARGET_ROOTFS}/imx${2}_rootfs${3}/usr/lib/
+ if [ $deploy_target_rd -eq 1 ]; then
 sudo cp -a $GPU_DIR/build/sdk/drivers/* ${TARGET_ROOTFS_RD}/imx${2}_rootfs${3}/usr/lib/
+  fi
 fi
 return $old_gpu_rc
 fi
 old_gpu_config=$1
+old_gpu_target=${2}${3}
 git branch -D build_${2}
 git checkout build || git add . && git commit -s -m"build $(date +%m%d)" && git checkout build
 git checkout -b build_${2} build
@@ -313,7 +348,9 @@ export VIVANTE_ENABLE_VG=1
 cd $AQROOT; make -j1 -f makefile.linux $BUILD_OPTIONS clean
 cd $AQROOT; make -j1 -f makefile.linux $BUILD_OPTIONS install 2>&1
 sudo cp -a $GPU_DIR/driver/build/sdk/drivers/* ${TARGET_ROOTFS}/imx${2}_rootfs${3}/usr/lib/
+ if [ $deploy_target_rd -eq 1 ]; then
 sudo cp -a $GPU_DIR/driver/build/sdk/drivers/* ${TARGET_ROOTFS_RD}/imx${2}_rootfs${3}/usr/lib/
+ fi
 old_gpu_rc=0
 return 0
 }
@@ -324,18 +361,26 @@ echo "make Platform $2 with $1"
 cd $KERNEL_DIR
 if [ "$old_kernel_config" = $1 ];then
 if [ "$old_kernel_rc" -eq 0 ]; then
+  if [ "$old_kernel_target" = "${2}${3}"  ]; then
+	return $old_kernel_rc
+  fi
 #sudo rm -rf ${TARGET_ROOTFS}/imx${2}_rootfs/lib/modules/*-daily*
 sudo make ARCH=arm modules_install INSTALL_MOD_PATH=${TARGET_ROOTFS}/imx${2}_rootfs${3} || return 3
+ if [ $deploy_target_rd -eq 1 ]; then
 sudo make ARCH=arm modules_install INSTALL_MOD_PATH=${TARGET_ROOTFS_RD}/imx${2}_rootfs${3} || return 3
+  fi
 scp arch/arm/boot/uImage root@10.192.225.218:/tftpboot/uImage_mx${2}_${3}d
 scp arch/arm/boot/uImage root@10.192.225.218:/var/ftp/uImage_mx${2}_${3}d
 scp arch/arm/boot/uImage ubuntu@10.192.244.7:/var/lib/tftpboot/uImage_mx${2}_${3}d
 sudo cp $KERNEL_DIR/tools/perf/perf ${TARGET_ROOTFS}/imx${2}_rootfs${3}/usr/bin/
+ if [ $deploy_target_rd -eq 1 ]; then
 sudo cp $KERNEL_DIR/tools/perf/perf ${TARGET_ROOTFS_RD}/imx${2}_rootfs${3}/usr/bin/
+  fi
 fi
 return $old_kernel_rc
 fi
 old_kernel_config=$1
+old_kernel_target=${2}${3}
 git branch -D build_${2}
 git checkout build || git add . && git commit -s -m"build $(date +%m%d)" && git checkout build
 git checkout -b build_${2} build
@@ -347,16 +392,20 @@ KERNEL_VER=$(./scripts/setlocalversion)
 #sudo rm -rf ${TARGET_ROOTFS}/imx${2}_rootfs${3}/lib/modules/*-daily*
 make ARCH=arm CROSS_COMPILE=${TOOL_CHAIN}arm-none-linux-gnueabi- -j 2 modules|| return 4
 sudo make ARCH=arm modules_install INSTALL_MOD_PATH=${TARGET_ROOTFS}/imx${2}_rootfs${3} || return 3
-sudo make ARCH=arm modules_install INSTALL_MOD_PATH=${TARGET_ROOTFS_RD}/imx${2}_rootfs${3} || return 3
 sudo make ARCH=arm headers_install INSTALL_HDR_PATH=${TARGET_ROOTFS}/imx${2}_rootfs${3}/usr/src/linux/ || return 5
+ if [ $deploy_target_rd -eq 1 ]; then
+sudo make ARCH=arm modules_install INSTALL_MOD_PATH=${TARGET_ROOTFS_RD}/imx${2}_rootfs${3} || return 3
 sudo make ARCH=arm headers_install INSTALL_HDR_PATH=${TARGET_ROOTFS_RD}/imx${2}_rootfs${3}/usr/src/linux/ || return 5
+ fi
 scp arch/arm/boot/uImage root@10.192.225.218:/tftpboot/uImage_mx${2}_${3}d
 scp arch/arm/boot/uImage root@10.192.225.218:/var/ftp/uImage_mx${2}_${3}d
 scp arch/arm/boot/uImage ubuntu@10.192.244.7:/var/lib/tftpboot/uImage_mx${2}_${3}d
 cd $KERNEL_DIR/tools/perf/
 make ARCH=arm CROSS_COMPILE=${TOOL_CHAIN}arm-none-linux-gnueabi- CFLAGS="-static -DGElf_Nhdr=Elf32_Nhdr"
 sudo cp  perf ${TARGET_ROOTFS}/imx${2}_rootfs${3}/usr/bin/
-sudo cp  perf ${TARGET_ROOTFS_RD}/imx${2}_rootfs${3}/usr/bin/
+ if [ $deploy_target_rd -eq 1 ]; then
+  sudo cp  perf ${TARGET_ROOTFS_RD}/imx${2}_rootfs${3}/usr/bin/
+  fi 
 old_kernel_rc=0
 return 0
 }
@@ -377,9 +426,11 @@ if [ "$old_vte_config" = $1 ]; then
    sudo cp -a install/* ${VTE_TARGET_PRE2}/vte_mx${2}_${3}d/
    sudo cp -a testcases/bin/* ${VTE_TARGET_PRE2}/vte_mx${2}_${3}d/testcases/bin/
    sudo cp mytest ${VTE_TARGET_PRE2}/vte_mx${2}_${3}d/
+   if [ $deploy_vte_target_rd -eq 1 ]; then
    sudo cp -a install/* ${VTE_TARGET_PRE3}/vte_mx${2}_${3}d/
    sudo cp -a testcases/bin/* ${VTE_TARGET_PRE3}/vte_mx${2}_${3}d/testcases/bin/
    sudo cp mytest ${VTE_TARGET_PRE3}/vte_mx${2}_${3}d/
+   fi
  fi
 return $old_vte_rc
 fi
@@ -418,9 +469,11 @@ sudo cp mytest ${VTE_TARGET_PRE}/vte_mx${2}_${3}d/
 sudo cp -a install/* ${VTE_TARGET_PRE2}/vte_mx${2}_${3}d/
 sudo cp -a testcases/bin/* ${VTE_TARGET_PRE2}/vte_mx${2}_${3}d/testcases/bin/
 sudo cp mytest ${VTE_TARGET_PRE2}/vte_mx${2}_${3}d/
+if [ $deploy_vte_target_rd -eq 1 ]; then
 sudo cp -a install/* ${VTE_TARGET_PRE3}/vte_mx${2}_${3}d/
 sudo cp -a testcases/bin/* ${VTE_TARGET_PRE3}/vte_mx${2}_${3}d/testcases/bin/
 sudo cp mytest ${VTE_TARGET_PRE3}/vte_mx${2}_${3}d/
+fi
 #sudo scp -r testcases/bin/* b17931@survivor:/rootfs/wb/vte_mx${2}_d/testcases/bin
 old_vte_rc=0
 return $ret
@@ -438,7 +491,9 @@ update_rootfs()
 {	
  cd $SCRPTSDIR
  sudo cp vte ${TARGET_ROOTFS}/imx${1}_rootfs${2}/etc/rc.d/init.d/vte 	
+ if [ $deploy_target_rd -eq 1 ]; then
  sudo cp vte_rd ${TARGET_ROOTFS_RD}/imx${1}_rootfs${2}/etc/rc.d/init.d/vte 	
+ fi
  deploy_firmware $1 $2
 }
 
@@ -637,7 +692,11 @@ do
   sync_server ${i} NOREADY
   while [ $j -lt $SOC_CNT ];do
    c_plat=${plat_name[${j}]}
+   deploy_target_rd=0
+   deploy_vte_target_rd=0
    if [ "$c_plat" = $i ];then
+     deploy_target_rd=${target_configs[${j}]}
+     deploy_vte_target_rd=${vte_target_configs[${j}]}
      c_soc=${soc_name[${j}]}
      make_target_tools MX${c_soc} 
      make_uboot ${u_boot_configs[${j}]} $c_soc $c_plat ${rootfs_apd[${j}]} || RC=$(echo $RC uboot_$i)
